+++ passes/S025/testdata/src/a/outside_package.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		AtLeastOneOf: []string{"test"},
+		Computed:     true,
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			AtLeastOneOf: []string{"test"},
+			Computed:     true,
+		},
+	}
+}