+++ passes/S025/testdata/src/a/schema/schema.go
@@ -1,0 +1,6 @@
+package schema
+
+type Schema struct {
+	AtLeastOneOf []string
+	Computed     bool
+}