+++ passes/S026/S026_test.go
@@ -1,0 +1,13 @@
+package S026_test
+
+import (
+	"testing"
+
+	"github.com/bflad/tfproviderlint/passes/S026"
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestS026(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, S026.Analyzer, "a")
+}