+++ passes/S027/testdata/src/a/comment_ignore.go
@@ -1,0 +1,13 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S027
+	_ = schema.Schema{
+		Computed: true,
+		Default:  "test",
+	}
+}