+++ passes/S032/testdata/src/a/outside_package.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		Computed: true,
+		MinItems: 1,
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			Computed: true,
+			MinItems: 1,
+		},
+	}
+}